refactoring according to psr-12
