+++ composer.json
@@ -8,7 +8,7 @@
     ],
     "autoload":{
         "psr-4":{
-            "app\\" : "./"
+            "App\\" : "./"
         }
     },
     "require-dev": {
