+++ core/LoginHelper.php
@@ -2,11 +2,11 @@
 
 declare(strict_types=1);
 
-namespace app\core;
+namespace App\core;
 
-use app\core\Application;
-use app\core\db\DBModel;
-use app\core\exception\ValidationException;
+use App\core\Application;
+use App\core\db\DBModel;
+use App\core\exception\ValidationException;
 
 class LoginHelper
 {
@@ -55,7 +55,7 @@ class LoginHelper
         $app->$class = $model;
         $primaryKey = $model->primaryKey();
         $primaryValue = $model->$primaryKey;
-        
+
         $app->session->set($class, $primaryValue);
         return true;
     }
