+++ core/Model.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace app\core;
+namespace App\core;
 
 abstract class Model
 {
@@ -17,7 +17,7 @@ abstract class Model
             if (!property_exists($this, $key)) {
                 return false;
             }
-            if (! $this->$key = $value){
+            if (!$this->$key = $value) {
                 return false;
             };
         }
