+++ core/Session.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace app\core;
+namespace App\core;
 
 class Session
 {
