+++ core/exception/ContainerException.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace app\core\exception;
+namespace App\core\exception;
 
 use Exception;
