+++ models/Post.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace app\models;
+namespace App\models;
 
-use app\core\db\DBModel;
+use App\core\db\DBModel;
 
 class Post extends DBModel
 {
