+++ public/index.php
@@ -4,16 +4,18 @@ declare(strict_types=1);
 
 require_once __DIR__ . "/../vendor/autoload.php";
 
-use app\core\Application;
+use App\core\Application;
 
 $envPath = "../.env";
 require "../LoadEnv.php";
 
 $config = [
+    // Define classes for authentication
     "class" => [
-        "userClass" => \app\models\User::class,
-        "adminClass" => \app\models\Admin::class,
+        "userClass" => \App\models\User::class,
+        "adminClass" => \App\models\Admin::class,
     ],
+    // Database Credentials
     "db" => [
         "dsn" => $data["DB_DSN"],
         "user" => $data["DB_USER"],
@@ -21,16 +23,18 @@ $config = [
     ]
 ];
 
-//set application dependencies
-include_once "../core/appServiceContainer.php";
+//set Application dependencies
+include_once "../core/AppServiceProvider.php";
 
 /**
- * Create new application instance
+ * Create new Application instance
  */
 $app = new Application(dirname(__DIR__), $config);
 
+// User routes
 include_once "../routes/user.php";
 
+// Admin routes
 include_once "../routes/admin.php";
 
 //runs the resolve method from Router.php after all requests are done
